refactor(images): extract client-safe variant tier constants
Move the variant tier ladder (VARIANT_TIER_WIDTHS), format helpers
(VARIANT_FORMATS, variantContentType), key builder (buildVariantKey) and
tierWidthsForSource into a new client-safe module ~/lib/image/variant-tiers
with no server-only dependencies.

server/lib/image-variants.ts re-exports them (behavior-preserving), so the
preprocessing pipeline keeps its existing import surface. The client-side
gallery loader (lib/image/loader.ts) can now import the same module instead
of mirroring the constants, eliminating the duplicated source of truth that
FE-3 had to maintain by hand.
